Export csv  and excel

+++ src/Controller/EnvironmentalConditionsController.php
@@ -4,6 +4,7 @@ namespace App\Controller;
 
 use App\Entity\CountingCampaign;
 use App\Entity\SiteCollection;
+use App\Entity\User;
 use App\Service\CampaignStatusService;
 use Symfony\Component\Security\Http\Attribute\IsGranted;
 use App\Entity\EnvironmentalConditions;
@@ -15,6 +16,7 @@ use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Attribute\Route;
+use Symfony\Contracts\Translation\TranslatorInterface;
 
 #[Route('/user/environmental/conditions')]
 #[IsGranted('ROLE_COLLECTOR', message: 'Vous n\'avez pas l\'accès.')]
@@ -43,105 +45,41 @@ class EnvironmentalConditionsController extends AbstractController
         ]);
     }
 
-    // #[Route('/new', name: 'app_environmental_conditions_new', methods: ['GET', 'POST'])]
-    // public function new(Request $request, EntityManagerInterface $entityManager): Response
-    // {
-    //     $user = $this->getUser(); // Récupérer l'utilisateur actuel
-    //     // Récupérer les identifiants du site et de la campagne depuis les paramètres de la requête
-    //     $siteId = $request->query->get('siteId');
-    //     $campaignId = $request->query->get('campaignId');
-
-    //     // Récupérer les entités Site et Campaign associées
-    //     $site = $entityManager->getRepository(SiteCollection::class)->find($siteId);
-    //     $campaign = $entityManager->getRepository(CountingCampaign::class)->find($campaignId);
-        
-    //     if (!$site || !$campaign) {
-    //         throw $this->createNotFoundException('Site or Campaign not found');
-    //     }
-        
-    //     if ($user) {
-    //         $this->addFlash('error', "L'utilsateur a deja cree conditions d'environnement pour cette site");
-    //     }
-    //     $environmentalCondition = new EnvironmentalConditions();
-    //     $form = $this->createForm(EnvironmentalConditionsType::class, $environmentalCondition);
-    //     $form->handleRequest($request);
-
-    //     if ($form->isSubmitted()) {
-    //         if ($form->isValid()) {
-    //             $environmentalCondition->setCreatedAt(new \DateTimeImmutable());
-    //             $environmentalCondition->setSiteCollection($site);
-    //             $environmentalCondition->setCountingCampaign($campaign);
-    //             $environmentalCondition->setUser($user);
-    //             $entityManager->persist($environmentalCondition);
-    //             $entityManager->flush();
-    //             $this->addFlash('success', "Conditions d'environnement a bien été crée");
-    
-    //             return $this->redirectToRoute('app_collected_data_new', [], Response::HTTP_SEE_OTHER);
-    //         } else {
-    //             // Log the errors for debugging
-    //             foreach ($form->getErrors(true) as $error) {
-    //                 error_log($error->getMessage());
-    //             }
-    //             $this->addFlash('error','Une erreur s\'est produite lors de la création de conditions d\'environnement.');
-    //         }
-    //     }
-
-    //     return $this->render('environmental_conditions/new.html.twig', [
-    //         'environmental_condition' => $environmentalCondition,
-    //         'form' => $form,
-    //     ]);
-    // }
-
 
     #[Route('/new', name: 'app_environmental_conditions_new', methods: ['GET', 'POST'])]
-    public function new(Request $request, EntityManagerInterface $entityManager): Response
+    public function new(Request $request, EntityManagerInterface $entityManager, TranslatorInterface $translator): Response
     {
         $user = $this->getUser(); // Récupérer l'utilisateur actuel
         $siteId = $request->query->get('siteId');
         $campaignId = $request->query->get('campaignId');
 
-        $statuses =$this->campaignStatusService->getCampaignStatuses();
-        $statusClosed = $statuses['statusClosed'];
-        $statusSuspended = $statuses['statusSuspended'];
-        $statusCancelled = $statuses['statusCancelled'];
-
         $site = $entityManager->getRepository(SiteCollection::class)->find($siteId);
         $campaign = $entityManager->getRepository(CountingCampaign::class)->find($campaignId);
         
+        // Vérifier si la campagne et le site existent
         if (!$campaign) {
-            throw $this->createNotFoundException('Aucune campagne trouvée');
+            $this->addFlash('error', 'La campagne spécifiée est introuvable.');
+            return $this->redirectToRoute('app_counting_campaign_index');
         }
+        
         if (!$site) {
-            throw $this->createNotFoundException('Aucun site trouvé');
+            $this->addFlash('error', 'Le site spécifié est introuvable.');
+            return $this->redirectToRoute('app_counting_campaign_index');
         }
-
-        if ($campaign->getCampaignStatus() === $statusClosed) {
-            throw $this->createNotFoundException("Une campagne $statusClosed ne peut pas être modifiée");
-        }
-        if ($campaign->getCampaignStatus() === $statusSuspended) {
-            throw $this->createNotFoundException("Une campagne $statusSuspended ne peut pas être modifiée");
-        }
-        if ($campaign->getCampaignStatus() === $statusCancelled) {
-            throw $this->createNotFoundException("Une campagne $statusClosed ne peut pas être modifiée");
-        }
-
-        // Vérifier si l'utilisateur est membre d'un groupe d'agents assigné à ce site
-        $isMember = false;
-        foreach ($site->getSiteAgentsGroups() as $siteAgentsGroup) {
-            foreach ($siteAgentsGroup->getAgentsGroup() as $group) {
-                if ($group->getGroupMember()->contains($user)) {
-                    $isMember = true;
-                    break;
-                }
-            }
-            if ($isMember) break;
-        }
-
+        
         // Si l'utilisateur n'est pas membre d'un groupe, interdire l'accès
-        if (!$isMember) {
-            $this->addFlash('warning', 'Vous ne pouvez pas ajouter de conditions car vous n\'êtes pas membre d\'un groupe assigné à ce site.');
-            return $this->redirectToRoute('app_counting_campaign_show', ['id' => $campaign->getId()]);
+        if (!$this->isUserSiteMember($user, $site)) {
+            $this->addFlash('warning', $translator->trans('member_of_a_group_assigned_to_this_site'));
+            return $this->redirectToRoute('app_environmental_conditions_index', [], Response::HTTP_SEE_OTHER);
         }
+        
+        // Vérifier si la campagne est modifiable
+        if ($campaign && !$this->campaignStatusService->ensureCampaignIsEditable($campaign)) {
+            // La vérification échoue, un message flash est déjà ajouté par le service
+            $this->addFlash('error', $translator->trans('campaign.cannot_be_modified', ['%status%' => $campaign->getCampaignStatus()]));
+            return $this->redirectToRoute('app_counting_campaign_show', ['id' => $campaignId]);
+        }
+
 
         // Vérifiez si l'utilisateur a déjà créé une condition environnementale pour ce site et cette campagne
         $existingCondition = $entityManager->getRepository(EnvironmentalConditions::class)->findOneBy([
@@ -180,12 +118,31 @@ class EnvironmentalConditionsController extends AbstractController
     }
 
 
+    private function isUserSiteMember(User $user, SiteCollection $site): bool
+    {
+        foreach ($site->getSiteAgentsGroups() as $siteAgentsGroup) {
+            foreach ($siteAgentsGroup->getAgentsGroup() as $group) {
+                if ($group->getGroupMember()->contains($user)) {
+                    return true;
+                }
+            }
+        }
+        return false;
+    }
+
+
     #[Route('/{id}', name: 'app_environmental_conditions_show', methods: ['GET'])]
     public function show(EnvironmentalConditions $environmentalCondition): Response
     {
-        return $this->render('environmental_conditions/show.html.twig', [
-            'environmental_condition' => $environmentalCondition,
-        ]);
+        // Vérifiez si l'utilisateur est admin, créateur ou membre du groupe
+        if ($this->isGranted('ROLE_ADMIN') || $this->isGranted('ROLE_VIEW') || $environmentalCondition->getUser() === $this->getUser()) {
+            return $this->render('environmental_conditions/show.html.twig', [
+                'environmental_condition' => $environmentalCondition,
+            ]);
+        } else {
+            $this->addFlash('info', 'Vous n\'avez pas accès à cette condition environnementale.');
+            return $this->redirectToRoute('app_environmental_conditions_index');
+        }
     }
 
     #[Route('/{id}/edit', name: 'app_environmental_conditions_edit', methods: ['GET', 'POST'])]
@@ -199,7 +156,7 @@ class EnvironmentalConditionsController extends AbstractController
         $statusSuspended = $statuses['statusSuspended'];
         $statusCancelled = $statuses['statusCancelled'];
         
-        if ($user != $environmentalCondition->getUser() && !$this->isGranted('ROLE_ADMIN')) {
+        if ($user != $environmentalCondition->getUser() && !$this->isGranted('ROLE_ADMIN') && !$this->isGranted('ROLE_EDIT')) {
             $this->addFlash('error', "Vous n'avez pas l'autorisation pour faire la modification");
             return $this->redirectToRoute('app_environmental_conditions_index', [], Response::HTTP_SEE_OTHER);
         }
@@ -236,8 +193,12 @@ class EnvironmentalConditionsController extends AbstractController
     }
 
     #[Route('/{id}', name: 'app_environmental_conditions_delete', methods: ['POST'])]
-    public function delete(Request $request, EnvironmentalConditions $environmentalCondition, EntityManagerInterface $entityManager): Response
+    public function delete(Request $request, EnvironmentalConditions $environmentalCondition, EntityManagerInterface $entityManager, TranslatorInterface $translator): Response
     {
+        if ($this->getUser() !== $environmentalCondition->getUser() && !$this->isGranted('ROLE_DELETE')){
+            $this->addFlash('error', $translator->trans('delete_permission'));
+            return $this->redirectToRoute('app_environmental_conditions_index', [], Response::HTTP_SEE_OTHER);
+        }
         if ($this->isCsrfTokenValid('delete'.$environmentalCondition->getId(), $request->getPayload()->get('_token'))) {
             $entityManager->remove($environmentalCondition);
             $entityManager->flush();
